requirements.txt
flask==2.2.2
environs==9.5.0
gunicorn==20.1.0
pytesseract==0.3.10
Pillow==9.4.0